queue: make queue pools safer
The automatic sharing of queue pools, together with the queue API
returning `*T` pointers can lead to code inadvertently or maliciously
reading from/writing into the values from another queue (even from a
separate package, if the type T is exported).

We remove the singleton map and require the caller to pass a pool.